Fix according to review comments
