+++ lib/manageiq/messaging/received_message.rb
@@ -1,0 +1,11 @@
+module ManageIQ
+  module Messaging
+    class ReceivedMessage
+      attr_accessor :sender, :message, :payload, :ack_ref
+
+      def initialize(sender, message, payload, ack_ref)
+        @sender, @message, @payload, @ack_ref = sender, message, payload, ack_ref
+      end
+    end
+  end
+end
