+++ lib/manageiq/messaging/stomp/queue.rb
@@ -4,9 +4,9 @@ module ManageIQ
       module Queue
         def publish_message_impl(options, &block)
           address, headers = queue_for_publish(options)
-          headers[:sender] = options[:sender] if options[:sender]
-          headers[:message_type] = options[:message] if options[:message]
-          headers[:class_name] = options[:class_name] if options[:class_name]
+          headers[:sender]         = options[:sender] if options[:sender]
+          headers[:message_type]   = options[:message] if options[:message]
+          headers[:class_name]     = options[:class_name] if options[:class_name]
           headers[:correlation_id] = Time.now.to_i.to_s if block_given?
 
           raw_publish(address, options[:payload] || '', headers)
@@ -30,7 +30,7 @@ module ManageIQ
             message_body = decode_body(msg.headers, msg.body)
             logger.info("Message received: queue(#{queue_name}), msg(#{payload_log(message_body)}), headers(#{msg.headers})")
 
-            result = yield [Struct::ManageIQ_Messaging_ReceivedMessage.new(sender, message_type, message_body, msg)]
+            result = yield [ManageIQ::Messaging::ReceivedMessage.new(sender, message_type, message_body, msg)]
             logger.info("Message processed")
 
             correlation_ref = msg.headers['correlation_id']
