+++ lib/manageiq/messaging/stomp/topic.rb
@@ -17,14 +17,13 @@ module ManageIQ
 
           subscribe(queue_name, headers) do |event|
             ack(event)
-            begin
-              sender = event.headers['sender']
-              event_type = event.headers['event_type']
-              event_body = decode_body(event.headers, event.body)
-              logger.info("Event received: queue(#{queue_name}), event(#{payload_log(event_body)}), headers(#{event.headers})")
-              yield sender, event_type, event_body
-              logger.info("Event processed")
-            end
+
+            sender = event.headers['sender']
+            event_type = event.headers['event_type']
+            event_body = decode_body(event.headers, event.body)
+            logger.info("Event received: queue(#{queue_name}), event(#{event_body}), headers(#{event.headers})")
+            yield sender, event_type, event_body
+            logger.info("Event processed")
           end
         end
       end
